Unit tests working
