+++ nomic/__init__.py
@@ -1,3 +1,3 @@
-from .project import AtlasProject
 from .embedders import CohereEmbedder
+from .project import AtlasProject
 from .projection import AtlasProjection
